pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "confluence-export-cli"
version = "0.1.0"
description = "CLI for exporting Confluence content for AI analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.scripts]
confluence-export = "confluence_cli.cli:app"

[tool.setuptools]
packages = ["confluence_cli"]
